generator/requirements.txt
nodemon
sqlalchemy
psycopg2-binary
watchdog
py-sudoku
